A story of  a little boy
It a small story of a boy who is have his normal life. As others might have in his age. The story is not completed. The thing which are pending are the name, and some other details. Which the user can set according to there wish. The steps to set is given below:
It is basically like x value and y value.